+++ src/normalizer/normalizer.ts
@@ -5,12 +5,12 @@ import {Dictionary} from '../types';
  * @param str - String to be normalized
  * @param charMap - CharMap used for normalization
  */
-export function normalize(str: string, charMap: Dictionary<string>)
+export function normalize(str: string, charMap: Dictionary<string>): string
 {
     Object.keys(charMap).forEach(chars =>
     {
-        let normalized = charMap[chars];
-        let regex = new RegExp('[' + chars + ']', 'gi');
+        const normalized = charMap[chars];
+        const regex = new RegExp('[' + chars + ']', 'gi');
 
         str = str.replace(regex, normalized);
     });