+++ src/types/enums.ts
@@ -1,0 +1,4 @@
+/**
+ * Type that represents enum type
+ */
+export type Enum<E = unknown> = Record<keyof E, number | string> & {[k: number]: string};