+++ src/types/index.ts
@@ -1,2 +1,4 @@
 export * from './dictionaries';
+export * from './enums';
+export * from './functions';
 export * from './valueNamePair';