+++ src/utils/nameof.ts
@@ -2,4 +2,4 @@
  * Returns name of property, checked by typescript
  * @param name - Name of property that should be returned
  */
-export const nameof = <T>(name: keyof T) => name;
+export const nameof = <T>(name: Extract<keyof T, string>): string => name;